requirements.txt
pandas==2.0.3
scikit-learn==1.3.0
fastapi==0.103.1
uvicorn==0.23.2
streamlit==1.26.0
pyyaml==6.0.1
requests==2.31.0
joblib==1.3.2

# 如果你要用 GCS 或 S3，解开下面对应的包

google-cloud-storage

boto3

mlflow

wandb